requirements.txt
requests
django==2.2.6
pytest
pytest-pythonpath==0.7.3
pytest-django
djangorestframework
djangorestframework-simplejwt
pillow
django-filter
djoser
djangorestframework_simplejwt 
